fix some stylistic issues
