+++ source/content.js
@@ -76,7 +76,7 @@ import hideNavigationHoverHighlight from './features/hide-navigation-hover-highl
 import displayIssueSuggestions from './features/display-issue-suggestions';
 import addPullRequestHotkey from './features/add-pull-request-hotkey';
 import openSelectionInNewTab from './features/add-selection-in-new-tab';
-import addClosedMilestoneDueDate from './features/add-closed-milestone-dueDate';
+import addClosedMilestoneDueDate from './features/add-closed-milestone-due-date';
 
 import * as pageDetect from './libs/page-detect';
 import {safeElementReady, enableFeature, safeOnAjaxedPages, injectCustomCSS} from './libs/utils';
@@ -200,7 +200,7 @@ function ajaxedPagesHandler() {
 	if (pageDetect.isMilestone()) {
 		enableFeature(addMilestoneNavigation); // Needs to be before sortMilestonesByClosestDueDate
 	}
-	
+
 	if (pageDetect.isClosedMilestoneList()) {
 		enableFeature(addClosedMilestoneDueDate);
 	}
